requirements.txt
asgiref==3.3.1
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.1
django-cpf==0.1.0
django-crispy-forms==1.11.2
django-heroku==0.3.1
django-import-export==2.5.0
django-input-mask==2.0.2
et-xmlfile==1.0.1
gunicorn==20.0.4
Jinja2==2.11.3
MarkupPy==1.14
MarkupSafe==1.1.1
odfpy==1.4.1
openpyxl==3.0.7
psycopg2==2.8.6
python-stdnum==1.16
pytz==2021.1
PyYAML==5.4.1
six==1.16.0
sqlparse==0.4.1
tablib==3.0.0
whitenoise==5.2.0
xlrd==2.0.1
xlwt==1.3.0
